oh yeah, i forgot about floor. oops.

+++ src/com/catsofwar/ittyblitty/Col.java
@@ -29,9 +29,7 @@
 
 package com.catsofwar.ittyblitty;
 
-import static java.lang.Math.max;
-import static java.lang.Math.min;
-import static util.Mth.floor;
+import static java.lang.Math.*;
 
 public class Col
 {
@@ -109,8 +107,8 @@ public class Col
 		}
 	
 		final float r, g, b;
-		float hsect = (h - floor(h)) * 6;
-		float hfrac = hsect - floor(hsect);
+		float hsect = (int)(h - floor((h)) * 6);
+		float hfrac = (int)(hsect - floor(hsect));
 		float q1 = v * (1 - s);
 		float q2 = v * (1 - s * hfrac);
 		float q3 = v * (1 - (s * (1 - hfrac)));
